+++ bin/install_uninstall.mli
@@ -1,0 +1,3 @@
+val install : unit Cmdliner.Term.t * Cmdliner.Term.info
+
+val uninstall : unit Cmdliner.Term.t * Cmdliner.Term.info