+++ src/dune/file_tree.ml
@@ -35,33 +35,21 @@ module Dune_file = struct
       }
   end
 
-  module Contents = struct
-    type t =
-      | Plain of Plain.t
-      | Ocaml_script of Path.Source.t
-  end
-
   type t =
-    { contents : Contents.t
-    ; kind : Dune_lang.File_syntax.t
-    }
+    | Plain of Plain.t
+    | Ocaml_script of Path.Source.t
 
-  let path t =
-    match t.contents with
+  let path = function
     | Plain x -> x.path
     | Ocaml_script p -> p
 
-  let load file ~project ~kind =
+  let load file ~project =
     Io.with_lexbuf_from_file (Path.source file) ~f:(fun lb ->
-        let contents, sub_dirs =
+        let t, sub_dirs =
           if Dune_lexer.is_script lb then
-            (Contents.Ocaml_script file, Sub_dirs.default)
+            (Ocaml_script file, Sub_dirs.default)
           else
-            let sexps =
-              Dune_lang.Parser.parse lb
-                ~lexer:(Dune_lang.Lexer.of_syntax kind)
-                ~mode:Many
-            in
+            let sexps = Dune_lang.Parser.parse lb ~mode:Many in
             let decoder =
               Dune_project.set_parsing_context project Sub_dirs.decode
             in
@@ -71,7 +59,7 @@ module Dune_file = struct
             in
             (Plain { path = file; sexps }, sub_dirs)
         in
-        ({ contents; kind }, sub_dirs))
+        (t, sub_dirs))
 end
 
 module Dir = struct
@@ -205,22 +193,22 @@ let readdir path =
     }
     |> Result.ok
 
-let load ?(warn_when_seeing_jbuild_file = true) path ~ancestor_vcs =
+let load path ~ancestor_vcs ~recognize_jbuilder_projects =
   let open Result.O in
   let nb_path_visited = ref 0 in
   let rec walk path ~dirs_visited ~project:parent_project ~vcs
-      ~(dir_status : Sub_dirs.Status.t) : (_, _) Result.t =
+      ~(dir_status : Sub_dirs.Status.t) { dirs; files } =
     incr nb_path_visited;
     if !nb_path_visited mod 100 = 0 then
       Console.update_status_line
         (Pp.verbatim (Printf.sprintf "Scanned %i directories" !nb_path_visited));
-    let+ { dirs; files } = readdir path in
     let project =
       if dir_status = Data_only then
         parent_project
       else
         Option.value
-          (Dune_project.load ~dir:path ~files)
+          (Dune_project.load ~dir:path ~files
+             ~infer_from_opam_files:recognize_jbuilder_projects)
           ~default:parent_project
     in
     let vcs =
@@ -242,57 +230,33 @@ let load ?(warn_when_seeing_jbuild_file = true) path ~ancestor_vcs =
         (let dune_file, sub_dirs =
            if dir_status = Data_only then
              (None, Sub_dirs.default)
-           else
-             let dune_file, sub_dirs =
-               match
-                 List.filter [ "dune"; "jbuild" ] ~f:(String.Set.mem files)
-               with
-               | [] -> (None, Sub_dirs.default)
-               | [ fn ] ->
-                 let file = Path.Source.relative path fn in
-                 let warn_about_jbuild =
-                   warn_when_seeing_jbuild_file && dir_status <> Vendored
-                 in
-                 if fn = "dune" then
-                   ignore
-                     ( Dune_project.ensure_project_file_exists project
-                       : Dune_project.created_or_already_exist )
-                 else if Dune_project.dune_version project >= (2, 0) then
-                   User_warning.emit
-                     ~loc:(Loc.in_file (Path.source file))
-                     [ Pp.text
-                         "jbuild files are not allowed inside Dune 2.0 \
-                          projects, please convert this file to a dune file \
-                          instead."
-                     ; Pp.text
-                         "Note: You can use \"dune upgrade\" to convert your \
-                          project to dune."
-                     ]
-                 else if warn_about_jbuild then
-                   User_warning.emit
-                     ~loc:(Loc.in_file (Path.source file))
-                     [ Pp.text
-                         "jbuild files are deprecated, please convert this \
-                          file to a dune file instead."
-                     ; Pp.text
-                         "Note: You can use \"dune upgrade\" to convert your \
-                          project to dune."
-                     ];
-                 let dune_file, sub_dirs =
-                   Dune_file.load file ~project
-                     ~kind:
-                       (Option.value_exn (Dune_lang.File_syntax.of_basename fn))
-                 in
-                 (Some dune_file, sub_dirs)
-               | _ ->
-                 User_error.raise
-                   [ Pp.textf
-                       "Directory %s has both a 'dune' and 'jbuild' file.\n\
-                        This is not allowed"
-                       (Path.Source.to_string_maybe_quoted path)
-                   ]
-             in
-             (dune_file, sub_dirs)
+           else (
+             if
+               (not recognize_jbuilder_projects)
+               && String.Set.mem files "jbuild"
+             then
+               User_error.raise
+                 ~loc:
+                   (Loc.in_file
+                      (Path.source (Path.Source.relative path "jbuild")))
+                 [ Pp.text
+                     "jbuild files are no longer supported, please convert \
+                      this file to a dune file instead."
+                 ; Pp.text
+                     "Note: You can use \"dune upgrade\" to convert your \
+                      project to dune."
+                 ];
+             if not (String.Set.mem files "dune") then
+               (None, Sub_dirs.default)
+             else (
+               ignore
+                 ( Dune_project.ensure_project_file_exists project
+                   : Dune_project.created_or_already_exist );
+               let file = Path.Source.relative path "dune" in
+               let dune_file, sub_dirs = Dune_file.load file ~project in
+               (Some dune_file, sub_dirs)
+             )
+           )
          in
          let sub_dirs =
            Sub_dirs.eval sub_dirs ~dirs:(List.map ~f:(fun (a, _, _) -> a) dirs)
@@ -332,7 +296,8 @@ let load ?(warn_when_seeing_jbuild_file = true) path ~ancestor_vcs =
                             ]
                     in
                     match
-                      walk path ~dirs_visited ~project ~dir_status ~vcs
+                      let+ x = readdir path in
+                      walk path ~dirs_visited ~project ~dir_status ~vcs x
                     with
                     | Ok dir -> String.Map.set acc fn dir
                     | Error _ -> acc ))
@@ -342,11 +307,17 @@ let load ?(warn_when_seeing_jbuild_file = true) path ~ancestor_vcs =
     Dir.create ~path ~contents ~status:dir_status ~project ~vcs
   in
   let walk =
+    let+ x = readdir path in
+    let project =
+      match
+        Dune_project.load ~dir:path ~files:x.files ~infer_from_opam_files:true
+      with
+      | None -> Dune_project.anonymous ~dir:path
+      | Some p -> p
+    in
     walk path
       ~dirs_visited:(File.Map.singleton (File.of_source_path path) path)
-      ~dir_status:Normal
-      ~project:(Lazy.force Dune_project.anonymous)
-      ~vcs:ancestor_vcs
+      ~dir_status:Normal ~project ~vcs:ancestor_vcs x
   in
   Console.clear_status_line ();
   match walk with