+++ test/blackbox-tests/test-cases/github1426/run.t
@@ -1,7 +1,6 @@
-Changing permissions of a dependency doesn't cause re-run
+Changing write permissions of a dependency doesn't cause a re-run
   $ chmod -w ./script.sh
   $ dune build a_target --display short
         script a_target
   $ chmod +w ./script.sh
   $ dune build a_target --display short
-        script a_target