+++ CHANGES.md
@@ -117,6 +117,12 @@
   `%{lib:lib:..}` when the library does not exist. (#2597, fix #1541,
   @rgrinberg)
 
+- Add `--sections` option to `$ dune install` to install subsections of .install
+  files. This is useful for installing only the binaries in a workspace for
+  example. (#2609, fixes #2554, @rgrinberg)
+
+- Drop support for `jbuild` and `jbuild-ignore` files (#2607, @diml)
+
 1.11.3 (23/08/2019)
 -------------------